Trying to fix winner

+++ runoff.c
@@ -162,7 +162,7 @@ bool print_winner(void)
 {
     for (int i  = 0; i < candidate_count; i++)
     {
-        float winner_test = (voter_count / 2);
+        float winner_test = ((float)voter_count / 2);
         if (candidates[i].votes > winner_test)
         {
             printf("%s\n", candidates[i].name);
